Remove playground import/replace action
New designs related to updating business networks as fabric chaincode
make the current editor Import/Replace action redundant

Includes clean up of deploy code to a purely deploy module now that
import is no longer involved

Contributes to #3241
