+++ packages/composer-playground/.istanbul.yml
@@ -8,5 +8,5 @@ check:
     global:
         statements: 99.51
         branches: 97.19
-        functions: 98.89
+        functions: 98.78
         lines: 99.56
