+++ packages/composer-playground/src/app/editor/editor.component.html
@@ -36,14 +36,6 @@
     </div>
     <div class="actions">
         <div>
-            <button id="editor_import" type="button" class="action" (click)="openImportModal()">
-                <svg class="standard-icon" aria-hidden="true">
-                    <use xlink:href="#icon-upload_32"></use>
-                </svg>
-                Import/Replace
-            </button>
-        </div>
-        <div>
             <button id="editor_export" type="button" class="action" (click)="exportBNA()">
                 <svg class="standard-icon" aria-hidden="true">
                     <use xlink:href="#icon-download_32"></use>
