+++ packages/composer-playground/src/app/editor/editor.component.spec.ts
@@ -1011,163 +1011,6 @@ describe('EditorComponent', () => {
         });
     });
 
-    describe('openImportModal', () => {
-        it('should open the import modal', fakeAsync(() => {
-            let mockUpdatePackage = sinon.stub(component, 'updatePackageInfo');
-
-            mockFileService.loadFiles.returns([]);
-
-            let finishedImport = new BehaviorSubject<any>(true);
-
-            mockDrawer.open = sinon.stub().returns({
-                componentInstance: {
-                    finishedSampleImport: finishedImport
-                },
-                close: sinon.stub()
-            });
-
-            component.openImportModal();
-
-            finishedImport.next({deployed: true});
-
-            tick();
-
-            mockUpdatePackage.should.have.been.called;
-            mockFileService.loadFiles.should.have.been.called;
-        }));
-
-        it('should open the import modal and set file to readme', fakeAsync(() => {
-            let mockUpdatePackage = sinon.stub(component, 'updatePackageInfo');
-            let mockSetCurrentFile = sinon.stub(component, 'setCurrentFile');
-
-            let file = new EditorFile('myId', 'myDisplay', 'myContent', 'readme');
-
-            mockFileService.loadFiles.returns([file]);
-
-            let finishedImport = new BehaviorSubject<any>(true);
-
-            mockDrawer.open = sinon.stub().returns({
-                componentInstance: {
-                    finishedSampleImport: finishedImport
-                },
-                close: sinon.stub()
-            });
-
-            component.openImportModal();
-
-            finishedImport.next({deployed: true});
-
-            tick();
-
-            mockUpdatePackage.should.have.been.called;
-            mockFileService.loadFiles.should.have.been.called;
-            mockSetCurrentFile.should.have.been.calledWith(file);
-            mockAlertService.successStatus$.next.should.have.been.called;
-        }));
-
-        it('should open the import modal and set file to first one if no readme', fakeAsync(() => {
-            let mockUpdatePackage = sinon.stub(component, 'updatePackageInfo');
-            let mockSetCurrentFile = sinon.stub(component, 'setCurrentFile');
-
-            let file = new EditorFile('myId', 'myDisplay', 'myContent', 'model');
-
-            mockFileService.loadFiles.returns([file, new EditorFile('myId', 'myDisplay', 'myContent', 'script')]);
-
-            let finishedImport = new BehaviorSubject<any>(true);
-
-            mockDrawer.open = sinon.stub().returns({
-                componentInstance: {
-                    finishedSampleImport: finishedImport
-                },
-                close: sinon.stub()
-            });
-
-            component.openImportModal();
-
-            finishedImport.next({deployed: true});
-
-            tick();
-
-            mockUpdatePackage.should.have.been.called;
-            mockFileService.loadFiles.should.have.been.called;
-            mockSetCurrentFile.should.have.been.calledWith(file);
-            mockAlertService.successStatus$.next.should.have.been.called;
-        }));
-
-        it('should open the import modal and handle error', fakeAsync(() => {
-            let mockUpdatePackage = sinon.stub(component, 'updatePackageInfo');
-
-            let finishedImport = new BehaviorSubject<any>(true);
-
-            let drawerItem = {
-                componentInstance: {
-                    finishedSampleImport: finishedImport
-                },
-                close: sinon.stub()
-            };
-            mockDrawer.open = sinon.stub().returns(drawerItem);
-
-            component.openImportModal();
-
-            finishedImport.next({deployed: false, error: 'some error'});
-
-            tick();
-
-            mockUpdatePackage.should.not.have.been.called;
-            mockFileService.loadFiles.should.not.have.been.called;
-            drawerItem.close.should.have.been.called;
-            mockAlertService.errorStatus$.next.should.have.been.calledWith('some error');
-        }));
-
-        it('should open the import modal and handle cancel', fakeAsync(() => {
-            let mockUpdatePackage = sinon.stub(component, 'updatePackageInfo');
-
-            let finishedImport = new BehaviorSubject<any>(true);
-
-            let drawerItem = {
-                componentInstance: {
-                    finishedSampleImport: finishedImport
-                },
-                close: sinon.stub()
-            };
-            mockDrawer.open = sinon.stub().returns(drawerItem);
-
-            component.openImportModal();
-
-            finishedImport.next({deployed: false});
-
-            tick();
-
-            mockUpdatePackage.should.not.have.been.called;
-            mockFileService.loadFiles.should.not.have.been.called;
-            drawerItem.close.should.have.been.called;
-            mockAlertService.errorStatus$.next.should.not.have.been.called;
-        }));
-
-        it('should open the import modal, and handle cancel by escape without throwing an error', fakeAsync(() => {
-            let mockUpdatePackage = sinon.stub(component, 'updatePackageInfo');
-
-            let finishedImport = new BehaviorSubject<any>(true);
-
-            let drawerItem = {
-                componentInstance: {
-                    finishedSampleImport: finishedImport
-                },
-                close: sinon.stub()
-            };
-            mockDrawer.open = sinon.stub().returns(drawerItem);
-
-            component.openImportModal();
-
-            finishedImport.next({deployed: false, error: DrawerDismissReasons.ESC});
-
-            mockUpdatePackage.should.not.have.been.called;
-            mockFileService.loadFiles.should.not.have.been.called;
-            drawerItem.close.should.have.been.called;
-            mockAlertService.errorStatus$.next.should.not.have.been.called;
-        }));
-    });
-
     describe('exportBNA', () => {
 
         afterAll(() => {
