+++ packages/composer-playground/src/app/editor/editor.module.ts
@@ -25,11 +25,11 @@ import { EditorRoutingModule } from './editor-routing.module';
 import { AddFileComponent } from './add-file/add-file.component';
 import { FileImporterModule } from '../common/file-importer/file-importer.module';
 import { DirectivesModule } from '../directives/directives.module';
-import { ImportModule } from '../import/import.module';
+import { DeployModule } from '../deploy/deploy.module';
 import { FooterModule } from '../footer/footer.module';
 
 @NgModule({
-    imports: [CommonModule, FormsModule, NgbModule, PerfectScrollbarModule, CodemirrorModule, DirectivesModule, FileImporterModule, ImportModule, EditorRoutingModule, FooterModule],
+    imports: [CommonModule, FormsModule, NgbModule, PerfectScrollbarModule, CodemirrorModule, DirectivesModule, FileImporterModule, DeployModule, EditorRoutingModule, FooterModule],
     entryComponents: [AddFileComponent],
     declarations: [EditorComponent, EditorFileComponent, AddFileComponent]
 })
