+++ packages/composer-playground/src/app/services/samplebusinessnetwork.service.spec.ts
@@ -357,72 +357,6 @@ describe('SampleBusinessNetworkService', () => {
         })));
     });
 
-    describe('updateBusinessNetwork', () => {
-        let idCard;
-
-        beforeEach(() => {
-            idCard = new IdCard({userName: 'banana'}, {'x-type': 'web', 'name': 'myProfile'});
-            identityCardMock.getCurrentIdentityCard.returns(idCard);
-            identityCardMock.getCurrentCardRef.returns('myCardRef');
-        });
-
-        it('should update the business network definition', fakeAsync(inject([SampleBusinessNetworkService], (service: SampleBusinessNetworkService) => {
-            adminMock.update.returns(Promise.resolve());
-            adminMock.connect.returns(Promise.resolve());
-            clientMock.refresh.returns(Promise.resolve());
-            mockFileService.getBusinessNetworkName.returns('myNetwork');
-            mockFileService.getBusinessNetworkDescription.returns('myDescription');
-
-            let buildStub = sinon.stub(service, 'buildNetwork').returns({getName: sinon.stub().returns('newname')});
-
-            let metaData = {getPackageJson: sinon.stub().returns({})};
-            businessNetworkMock.getMetadata.returns(metaData);
-            businessNetworkMock.getMetadata.returns(metaData);
-
-            service.updateBusinessNetwork(businessNetworkMock);
-
-            tick();
-
-            metaData.getPackageJson.should.have.been.called;
-
-            buildStub.should.have.been.called;
-
-            adminMock.connect.should.have.been.calledWith('myCardRef', idCard, true);
-            adminMock.update.should.have.been.called;
-            clientMock.refresh.should.have.been.called;
-            adminMock.reset.should.have.been.calledWith('newname');
-            alertMock.busyStatus$.next.should.have.been.calledWith(null);
-        })));
-
-        it('should handle error', fakeAsync(inject([SampleBusinessNetworkService], (service: SampleBusinessNetworkService) => {
-            mockFileService.getBusinessNetworkName.returns('myNetwork');
-            mockFileService.getBusinessNetworkDescription.returns('myDescription');
-
-            let buildStub = sinon.stub(service, 'buildNetwork').returns({getName: sinon.stub()});
-
-            adminMock.connect.returns(Promise.resolve());
-            adminMock.update.returns(Promise.reject('some error'));
-
-            let metaData = {getPackageJson: sinon.stub().returns({})};
-            businessNetworkMock.getMetadata.returns(metaData);
-
-            service.updateBusinessNetwork(businessNetworkMock).then(() => {
-                throw('should not get here');
-            })
-                .catch((error) => {
-                    alertMock.busyStatus$.next.should.have.been.calledWith(null);
-                    error.should.equal('some error');
-                });
-
-            tick();
-
-            buildStub.should.have.been.called;
-
-            adminMock.connect.should.have.been.calledWith('myCardRef', idCard, true);
-            adminMock.update.should.have.been.called;
-        })));
-    });
-
     describe('buildNetwork', () => {
         it('should build the network', inject([SampleBusinessNetworkService], (service: SampleBusinessNetworkService) => {
             let modalManagerMock = {addModelFiles: sinon.stub(), getModelFiles: sinon.stub().returns(['model'])};
